feat(Date) : 최소 예약 취소 날짜를 넘어가는 경우 취소할 수 없다.
- 취소할 수 업다는 에러를 뱉는지 Test 작성
- 날짜를 제대로 확인하는지 Test 작성
- ReservationExceptionCode 추가

+++ src/main/java/roomescape/reservation/service/ReservationService.java
@@ -139,6 +139,10 @@ public class ReservationService {
         Reservation reservation = reservationRepository.findById(reservationId)
                 .orElseThrow(ReservationNotFoundException::new);
 
+        if (reservation.isAfterCancelDate(LocalDate.now())) {
+            throw new RoomEscapeException(ReservationExceptionCode.CAN_NOT_CANCEL_AFTER_MIN_CANCEL_DATE);
+        }
+
         reservationRepository.deleteById(reservationId);
         waitingToReservation(reservation);
     }
